sql: update max1row operator to take error text as a parameter
Previously, when the max1row operator found multiple input rows, it
raised an error with a literal string as text, that refers to subqueries.
This commit makes the error text a parameter of the max1row operator,
so that it can be used in other contexts besides subqueries. In
particular, it will soon be used for raising an error in the duplicate
row Upsert case.

Release note: None