opt: eliminate distinct operators with Values input
Add new rules that fold a distinct operator with a Values operator
containing constant grouping values. For DistinctOn, the rule will prune any
duplicate Values rows so that the distinct operation is no longer necessary.
By contrast, UpsertDistinctOn will only be eliminated if there are no
duplicate rows to start with. If duplicates are found, then the rule does not
match, and just lets execution raise an "UPSERT does not allow duplicates"
error. Here is an example of the DistinctOn case:

   SELECT DISTINCT (x) * FROM (VALUES (1), (1)) t(x)
   =>
   SELECT * FROM (VALUES (1)) t(x)

These rules are especially useful in the UpsertDistinctOn case, since every
UPSERT or INSERT..ON CONFLICT statement will generate an UpsertDistinctOn,
and it's common for it to wrap a VALUES input.

Release note: None